feat: Update migrate_teams.py to retain parent linkage.
Previously we were not retaining parent child relationships between
teams. Now we grab the parent info for a team on the fly and ensure the
same links exist between the new teams in the destination org.

This is important because sometimes users get access to repos because
they are given access based on their team's parent team.