feat: set team permissions during repo migration (#18)
This adds support for assigning team permissions to repos that have
been transferred over. It makes the following changes to the
migrate_repos.py script:

* created option --permissions-file
* --skip-missing was changed to be --skip-missing-repos
* created option --skip-missing-teams
* --github-token can be used intead of the $GITHUB_TOKEN env var

The permissions file expects one of our export-{org}.json files. It
is optional because there are situations where we'll want to bring
over repos without team permissions (e.g. from edx-solutions). But
it is *strongly* recommended, so it will prompt the user for
confirmation if they don't specify one.

The script assumes that:

1. All teams with assigned repo permissions in the source org have
   matching counterparts in the destination org with the same slug.
   So if there's a "teaching-and-learning" team in the source org,
   there needs to be a matching "teaching-and-learning" team in the
   destination org.
2. All users with assigned repo permissions in the source org are
   also members of the destination org.

The script will create the team permissions in the new org to match
what they were in the old org.

If you specify a permissions file, all repos that are being
transferred must have entries in it. There is no way to bypass this
check–the script assumes that if entries are missing, it's because a
new repo was created recently and we don't have proper permissions
data for it in our export yet (and so we should redo the export).

This script does not explicitly create user permissions, since those
carry over automatically when doing the repo transfer. However we do
check to see if the users exist in the destination org, so that we
don't accidentally make a bunch of people outside collaborators without
meaning to.

By default, if any teams or users are missing from the destination org
the script will exit with an error. However there are situations where
we might want to skip missing teams. For instance, we're capturing
the admins of the "edx" org as a special team "edx-admin", so that we
can give them the appropriate access without making them admins of the
"openedx" org. This "edx-admin" team doesn't exist in the "edx" GitHub
org. So if we were to need to rollback and move repos back over to
edx from openedx, the script will complain that "edx-admin" is not a
team in the "edx" org. The --skip-missing-teams option was created to
address this case. It should just be used with caution.