feat: add option to resume transfer with a particular repo
Adds a --resume-from option to the migrate_repos.py script, allowing a
user to resume transferring after their previous transfer was
interrupted by some error. Transfers go through the repos in the order
that they are listed in the REPO_LIST_FILE.

There is special case handling so that you can resume from a repo in the
scenario where the repo itself has transferred over, but the permissions
code was not properly applied. If you use this already-transferred repo
as your value for --resume-from, it will detect that it's already moved,
and simply reapply the team permissions changes (which was probably what
you intended when you wanted to resume from it).