feat: add repo lists for phases 2-7
The naming convention for these files is {phase}-{source-org}.txt

These lists were generated from the following spreadsheet:

  https://docs.google.com/spreadsheets/d/1vxuwG9z7-iHXuIFUVrxR8Giq8kvDvMrLJE4FrrehJnc/edit#gid=0

Phases 0 and 1 had repos from both the edx org and the edx-solutions
org for testing purposes, but in this batch, only phase 4 pulls from
the edx-solutions org.

The four repos in phase 8 are considered special cases, and will be
moved over individually.