feat: add special "Phase 8" repos.
These are our special handle-with-care repos that we're listing under
"Phase 8", though it really means, "we're going to handle these one at a
time, after we've done all the more straightforward phases".

The edx-platform repo is tricky because of how central it is and the
CI/CD machinery around it. Codejail has very particular testing
requirements to validate the sandboxing security (we currently use
Jenkins for this). Paragon has a peculiar Netlify-based build system
that builds PRs with multiple configs for the edX-themed and unthemed
versions.