feat: Add a new script to remove unused teams.
This version does not take into account the recursive aspect of team
removal where we currently delete teams that are leaf teams but this
might result in new leaf teams.  We could make it handle that but it's
easy enough to just run the script a few times.

Especially after the first run which will probably remove two thirds of
the teams.