chore: Update export post empty team removal.
I ran the `remove_unused_teams.py` command and removed all fully empty
teams. I also communicated the removal of the teams with members but
nothing else and in the case of content-aurora, Matt Hughes updated it
to be used in the correct places since it's the most corrent name for
that team.