feat: Add a script to convert direct user access to team access.
We want to avoid users having direct access to repos since it's hard to
know why they have that access over time.  With a team, there is more
opportunity to document why the team has access.  This is the script I
used to do quickly do this work. It's temporary and will be removed
after it's done.