feat: Skip security forks.
Security forks are not like regular repositories in most of the ways we
care about.  In particular, you don't manage access to them via teams
the same way since they are linked to a security advisory.

For now we'll skip these for all checks but later it will probably make
sense to figure out a way for each check to evaluate a repo for whether
or not it should run on that repo, independent of whether or not it will
pass or fail if run.  This was great advice from Dave Ormsbee and we
can iterate on it futher as we add more checks.