feat: Match labels with different emoji and casing.
To reduce the chance of us creating lots of labels if we end up changing
emoji or casing in the future.  And in the case where some repos might
alredy have a `maintenance` label. Make the matching of labels ignore
emoji, leading whitespace, trailing whitespce and casing.

`:+1: :-1: Maintenance` will match any version of the word maintenance
with any emoji in it.