feat: add "good first issue" label to all GitHub repos
The label will be added to all existing repos. Note that this might
conflict with the recent "good first issue :tada:" default label that is
part of the openedx org. That "tada" label will probably have to be
renamed, and repos that already have it should migrate the issues to use
the new label.

Color is a sweet green. Alternative suggestions are welcome.

Close #51.