feat: Be able to tell if a repo is still empty.
Add a new function to see if a repository is completly empty(zero
commits/branches).  This is useful because when we create new repos,
they are sometimes fully empty and so certain checks shouldn't run on
them.

Update the branch protection rule portion of the CLA check to use the
new function instead of assuming the repo might be empty if certain
errors happen and use the new function to ensure we don't try to add
workflows to an empty repository.

We may want to revisit the latter point in the future.  There may be
some files we want to add even to empty repos but we'll need special
logic for that so leave it alone for now.