refactor: extract repo_checks labels out to YAML file
We are preparing make the EnsureLabels checks the one-and-only
authoritative place for org-level labels. As part of this, we're
extracting the list of labels out of the Python script and into
a YAML file.

For repos that are already up-to-date with repo_checks, this change
should be a no-op.

Additionally, this cleans up the EnsureLabels code a bit so that it'll
be easier to add label-description support in the next commit.

Part of:
https://openedx.atlassian.net/wiki/spaces/COMM/pages/3695214629/openedx+org-wide+labels#1.-Proposal%3A-Store-labels-in-one-central-place