feat: set label descriptions in repo_checks
repo_checks currently normalizes label names and colors. With this
commit, it will also normalize the label description.

In interest of minimizing the effect of running the EnsureLabels check:
For labels that generally have already-established descriptions, we use
those. For labels that don't, we just use !!null for now, with the plan
of adding descriptions in the near future.

Part of:
https://openedx.atlassian.net/wiki/spaces/COMM/pages/3695214629/openedx+org-wide+labels#2.-Proposal%3A-List-of-starter-labels